Factor out markdown and excerpt rendering
